refactor: рефакторит код проекта
* Переименовывает папку modules в blocks
* Создает папку utils, определяет в ней общие утилитарные функции и для
  работы с датой
* Переносит моковые данные в папку global
* Временно изменяет реализацию функции render для исключения ошибок
