+++ src/global/mock/const.js
@@ -1,0 +1,62 @@
+const MIN_DEST_DESCR_RANDOM_VALUE = 1;
+const MAX_DEST_DESCR_RANDOM_VALUE = 5;
+const MIN_PICTURE_COUNT = 0;
+const MAX_PICTURE_COUNT = 5;
+const MIN_SRC_PICTURE_RANDOM_VALUE = 1;
+const MAX_SRC_PICTURE_RANDOM_VALUE = 9;
+const MIN_OFFER_COUNT = 0;
+const MAX_OFFER_COUNT = 8;
+const MIN_OFFER_PRICE_VALUE = 200;
+const MAX_OFFER_PRICE_VALUE = 2000;
+const MIN_EVENT_BASE_PRICE = 1000;
+const MAX_EVENT_BASE_PRICE = 5000;
+
+const DESTINATION_DESCRIPTIONS = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget.',
+  'Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra.',
+  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
+  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.',
+  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.',
+  'Sed sed nisi sed augue convallis suscipit in sed felis.',
+  'Aliquam erat volutpat.',
+  'Nunc fermentum tortor ac porta dapibus.',
+  'In rutrum ac purus sit amet tempus.'
+];
+
+const DESTINATION_NAMES = [
+  'Moscow',
+  'Paris',
+  'London',
+  'Pskov',
+  'Omsk'
+];
+
+const EVENT_TYPES = [
+  'taxi',
+  'bus',
+  'train',
+  'ship',
+  'drive',
+  'flight',
+  'check-in',
+  'sightseeing',
+  'restaurant'
+];
+
+export {
+  MIN_DEST_DESCR_RANDOM_VALUE,
+  MAX_DEST_DESCR_RANDOM_VALUE,
+  MIN_PICTURE_COUNT,
+  MAX_PICTURE_COUNT,
+  MIN_SRC_PICTURE_RANDOM_VALUE,
+  MAX_SRC_PICTURE_RANDOM_VALUE,
+  MIN_OFFER_COUNT,
+  MAX_OFFER_COUNT,
+  MIN_OFFER_PRICE_VALUE,
+  MAX_OFFER_PRICE_VALUE,
+  MIN_EVENT_BASE_PRICE,
+  MAX_EVENT_BASE_PRICE,
+  DESTINATION_DESCRIPTIONS,
+  DESTINATION_NAMES,
+  EVENT_TYPES
+};
