+++ src/main.js
@@ -1,6 +1,6 @@
-import EventsModel from './modules/events/model/events-model.js';
-import TripMainPresenter from './modules/trip-main/presenter/trip-main-presenter.js';
-import EventsPresenter from './modules/events/presenter/events-presenter.js';
+import EventsModel from './blocks/events/model/events-model.js';
+import TripMainPresenter from './blocks/trip-main/presenter/trip-main-presenter.js';
+import EventsPresenter from './blocks/events/presenter/events-presenter.js';
 
 const eventsModel = new EventsModel();
 const tripMainPresenter = new TripMainPresenter();
